moving keysyms out to their own crate
