+++ crates/penrose_keysyms/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "penrose_keysyms"
+version = "0.1.0"
+authors = ["IDAM <innes.andersonmorrison@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+strum = { version = "0.19.2", features = ["derive"] }
+strum_macros = "0.19.2"
