+++ src/core/bindings.rs
@@ -7,6 +7,9 @@ use crate::{
     PenroseError, Result,
 };
 
+#[cfg(feature = "keysyms")]
+use penrose_keysyms::XKeySym;
+
 use std::{collections::HashMap, convert::TryFrom};
 
 use strum::EnumIter;
@@ -52,6 +55,28 @@ pub enum KeyPress {
     Left,
     /// Right
     Right,
+}
+
+#[cfg(feature = "keysyms")]
+impl TryFrom<XKeySym> for KeyPress {
+    type Error = PenroseError;
+
+    fn try_from(s: XKeySym) -> std::result::Result<KeyPress, PenroseError> {
+        Ok(match s {
+            XKeySym::XK_Return | XKeySym::XK_KP_Enter | XKeySym::XK_ISO_Enter => KeyPress::Return,
+            XKeySym::XK_Escape => KeyPress::Escape,
+            XKeySym::XK_Tab | XKeySym::XK_ISO_Left_Tab | XKeySym::XK_KP_Tab => KeyPress::Tab,
+            XKeySym::XK_BackSpace => KeyPress::Backspace,
+            XKeySym::XK_Delete | XKeySym::XK_KP_Delete => KeyPress::Delete,
+            XKeySym::XK_Page_Up | XKeySym::XK_KP_Page_Up => KeyPress::PageUp,
+            XKeySym::XK_Page_Down | XKeySym::XK_KP_Page_Down => KeyPress::PageDown,
+            XKeySym::XK_Up | XKeySym::XK_KP_Up => KeyPress::Up,
+            XKeySym::XK_Down | XKeySym::XK_KP_Down => KeyPress::Down,
+            XKeySym::XK_Left | XKeySym::XK_KP_Left => KeyPress::Left,
+            XKeySym::XK_Right | XKeySym::XK_KP_Right => KeyPress::Right,
+            s => KeyPress::Utf8(s.as_utf8_string()?),
+        })
+    }
 }
 
 /// A u16 X key-code bitmask
